backend: split job intake logic out of the jobs controllers into a new sync/jobs.ts file
